% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{choose_theme}
\alias{choose_theme}
\title{Choose a theme for your front end}
\usage{
choose_theme(th_type = "cerulean")
}
\description{
This function return the UI theme dependeing on the choice of a user
}
