% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_fct_all.R
\docType{data}
\name{link_github}
\alias{link_github}
\title{some global variables}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
link_github
}
\description{
some global variables
}
\keyword{datasets}
