% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlRdrift-package.R
\docType{package}
\name{mlRdrift-package}
\alias{mlRdrift}
\alias{mlRdrift-package}
\title{mlRdrift: Machine Learning in R Drift Analyst}
\description{
Tool to perform the drift analyst during the monitoring of a machine learning model in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seewe/mlRdrift}
  \item Report bugs at \url{https://github.com/seewe/mlRdrift/issues}
}

}
\author{
\strong{Maintainer}: Sergio Ewane Ebouele \email{info@dataforknow.com}

}
\keyword{internal}
