% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_fct_all.R
\name{same_variables}
\alias{same_variables}
\title{Detect common variables in two dataset}
\usage{
same_variables(df1, df2)
}
\arguments{
\item{df1}{the first dataset to use}

\item{df2}{The second dataset to use}
}
\value{
a Character vector containing all variables names in both df1 and df2
}
\description{
Detect common variables in two dataset
}
\examples{
same_variables(iris, iris)
same_variables(mtcars, mtcars)
}
