% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_fct_all.R
\name{skim_char}
\alias{skim_char}
\title{Skim a dataset and return only characters variables characteristics}
\usage{
skim_char(the_data)
}
\arguments{
\item{the_data}{Data on which the skim function will apply the description on character variables}
}
\value{
a data.frame object containing description of all character (factor, character or date) variable in the input data.
}
\description{
Skim a dataset and return only characters variables characteristics
}
\examples{
skim_char(iris)
skim_char(mtcars)

}
