% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_fct_all.R
\name{skim_num}
\alias{skim_num}
\title{Skim a dataset and return only numeric variables characteristics}
\usage{
skim_num(the_data)
}
\arguments{
\item{the_data}{Data on which the skim function will apply the description on numeric variables}
}
\value{
a data.frame object containing description of all numeric (doucle or integer) variable in the input data.
}
\description{
Skim a dataset and return only numeric variables characteristics
}
\examples{
skim_num(iris)
skim_num(mtcars)

}
