% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_fct_all.R
\name{var_names}
\alias{var_names}
\title{var_names}
\usage{
var_names(input_data)
}
\arguments{
\item{input_data}{The data frame from which user need to have the variables names}
}
\value{
a character vector of variables names
}
\description{
A function which return the names of all variables in a dataframe
}
\details{
Variables names of a data.frame
}
\examples{
var_names(iris)
var_names(mtcars)
}
